Points

Description

Provides methods to work with points in a 2-dimensional space. All coordinates are in single float precision.

Documentation

convexHull :: [Point] -> [Point]

Returns a subset of the given list which contains all points that form the convex hull. That is, no outer angle is smaller than 180°. The returned list is sorted clockwise, starting with the point with the greatest y coordinate. Returns an empty list if an empty list is given or all points are located on a line.

center :: [Point] -> Point

Returns the center of the given point cloud. The center defined by the intersection of the both diagonals of the smallest rectangle that enfolds all points.