Sort

Description

Provides a method to sort points in a 2-dimensional space around a given center.

Documentation

sort :: [Point] -> Point -> [Point]

Sorts the given points around a center by comparing their direction. Returns a list of points sorted counter-clockwise, starting from the x-axi in the lower left quadrant. If two points have the same direction, their magnitude is compared with smaller magnitudes taking precedence. The latter behavior is especially important in order to detect non-convex points easily when encountering two points with the same direction.