Vector

Description

Provides methods to work with 2-dimensional vectors.

Documentation

type Vector = (Float, Float)

A tuple of x and y which describe a vector in a 2-dimensional space.

angle :: Vector -> Vector -> Float

Calculates the angle between two vectors.

magn :: Vector -> Float

Returns the length of a vector.

diff :: Point -> Point -> Vector

Caculates the vector from one pointer to another.

dot :: Vector -> Vector -> Float

Returns the result of the dot multiplication of two vectors.

deg :: Float -> Float

Translates radians into degrees.

orient :: Point -> Point -> Float

Returns the orientation of the vector that goes from one point to another. The result is one of -1.0, 0.0 and 1.0, meaning clockwise, equal and counter-clockwise respectively.